refactor: change function types to arrow
